requirements.txt
flask>=3.0.0
rembg>=2.0.50
pillow>=10.0.0
numpy>=1.24.0
onnxruntime>=1.16.0
